+++ packages/node/src/behavior/context/server/LocalActorContext.ts
@@ -139,7 +139,10 @@ export namespace LocalActorContext {
     /**
      * {@link LocalActorContext} configuration options.
      */
-    export interface Options extends Omit<InteractionSettings, "transaction"> {
+    export interface Options
+        extends
+            Omit<InteractionSettings, "transaction">,
+            Pick<ValueSupervisor.SupervisionSettings, "clientPeerContext"> {
         lifetime?: Lifetime.Owner;
         command?: boolean;
         activity?: NodeActivity;